+++ src/lib.rs
@@ -18,6 +18,7 @@ pub use self::solc::combined_json::contract::Contract as SolcCombinedJsonContrac
 pub use self::solc::combined_json::CombinedJson as SolcCombinedJson;
 pub use self::solc::pipeline::Pipeline as SolcPipeline;
 pub use self::solc::standard_json::input::language::Language as SolcStandardJsonInputLanguage;
+pub use self::solc::standard_json::input::settings::optimizer::Optimizer as SolcStandardJsonInputSettingsOptimizer;
 pub use self::solc::standard_json::input::settings::selection::file::flag::Flag as SolcStandardJsonInputSettingsSelectionFileFlag;
 pub use self::solc::standard_json::input::settings::selection::file::File as SolcStandardJsonInputSettingsSelectionFile;
 pub use self::solc::standard_json::input::settings::selection::Selection as SolcStandardJsonInputSettingsSelection;
@@ -39,12 +40,10 @@ use std::path::PathBuf;
 pub fn yul(
     input_files: &[PathBuf],
     solc: &SolcCompiler,
-    optimize: bool,
+    optimizer_settings: compiler_llvm_context::OptimizerSettings,
     is_system_mode: bool,
     debug_config: Option<compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<Build> {
-    let solc_version = solc.version()?;
-
     let path = match input_files.len() {
         1 => input_files.first().expect("Always exists"),
         0 => anyhow::bail!("The input file is missing"),
@@ -54,13 +53,12 @@ pub fn yul(
         ),
     };
 
-    let project = Project::try_from_yul_path(path, &solc_version.default)?;
+    if !is_system_mode {
+        solc.yul(input_files)?;
+    }
 
-    let optimizer_settings = if optimize {
-        compiler_llvm_context::OptimizerSettings::size()
-    } else {
-        compiler_llvm_context::OptimizerSettings::none()
-    };
+    let project = Project::try_from_yul_path(path)?;
+
     let target_machine = compiler_llvm_context::TargetMachine::new(&optimizer_settings)?;
     let build = project.compile_all(
         target_machine,
@@ -77,7 +75,7 @@ pub fn yul(
 ///
 pub fn llvm_ir(
     input_files: &[PathBuf],
-    optimize: bool,
+    optimizer_settings: compiler_llvm_context::OptimizerSettings,
     is_system_mode: bool,
     debug_config: Option<compiler_llvm_context::DebugConfig>,
 ) -> anyhow::Result<Build> {
@@ -92,11 +90,6 @@ pub fn llvm_ir(
 
     let project = Project::try_from_llvm_ir_path(path)?;
 
-    let optimizer_settings = if optimize {
-        compiler_llvm_context::OptimizerSettings::size()
-    } else {
-        compiler_llvm_context::OptimizerSettings::none()
-    };
     let target_machine = compiler_llvm_context::TargetMachine::new(&optimizer_settings)?;
     let build = project.compile_all(
         target_machine,
@@ -116,7 +109,7 @@ pub fn standard_output(
     input_files: &[PathBuf],
     libraries: Vec<String>,
     solc: &SolcCompiler,
-    optimize: bool,
+    optimizer_settings: compiler_llvm_context::OptimizerSettings,
     force_evmla: bool,
     is_system_mode: bool,
     base_path: Option<String>,
@@ -132,8 +125,13 @@ pub fn standard_output(
         input_files,
         libraries,
         SolcStandardJsonInputSettingsSelection::new_required(solc_pipeline),
-        true,
+        SolcStandardJsonInputSettingsOptimizer::new(true, None),
     )?;
+    let source_code_files = solc_input
+        .sources
+        .iter()
+        .map(|(path, source)| (path.to_owned(), source.content.to_owned()))
+        .collect();
 
     let libraries = solc_input.settings.libraries.clone().unwrap_or_default();
     let mut solc_output = solc.standard_json(solc_input, base_path, include_paths, allow_paths)?;
@@ -155,17 +153,13 @@ pub fn standard_output(
     }
 
     let project = solc_output.try_to_project(
+        source_code_files,
         libraries,
         solc_pipeline,
         &solc_version.default,
         debug_config.as_ref(),
     )?;
 
-    let optimizer_settings = if optimize {
-        compiler_llvm_context::OptimizerSettings::size()
-    } else {
-        compiler_llvm_context::OptimizerSettings::none()
-    };
     let target_machine = compiler_llvm_context::TargetMachine::new(&optimizer_settings)?;
     let build = project.compile_all(
         target_machine,
@@ -196,7 +190,14 @@ pub fn standard_json(
     let zksolc_version = semver::Version::parse(env!("CARGO_PKG_VERSION")).expect("Always valid");
 
     let solc_input = SolcStandardJsonInput::try_from_stdin(solc_pipeline)?;
-    let optimize = solc_input.settings.optimizer.enabled;
+    let source_code_files = solc_input
+        .sources
+        .iter()
+        .map(|(path, source)| (path.to_owned(), source.content.to_owned()))
+        .collect();
+
+    let optimizer_settings =
+        compiler_llvm_context::OptimizerSettings::try_from(&solc_input.settings.optimizer)?;
 
     let libraries = solc_input.settings.libraries.clone().unwrap_or_default();
     let mut solc_output = solc.standard_json(solc_input, base_path, include_paths, allow_paths)?;
@@ -211,17 +212,13 @@ pub fn standard_json(
     }
 
     let project = solc_output.try_to_project(
+        source_code_files,
         libraries,
         solc_pipeline,
         &solc_version.default,
         debug_config.as_ref(),
     )?;
 
-    let optimizer_settings = if optimize {
-        compiler_llvm_context::OptimizerSettings::size()
-    } else {
-        compiler_llvm_context::OptimizerSettings::none()
-    };
     let target_machine = compiler_llvm_context::TargetMachine::new(&optimizer_settings)?;
     let build = project.compile_all(
         target_machine,
@@ -244,7 +241,7 @@ pub fn combined_json(
     input_files: &[PathBuf],
     libraries: Vec<String>,
     solc: &SolcCompiler,
-    optimize: bool,
+    optimizer_settings: compiler_llvm_context::OptimizerSettings,
     force_evmla: bool,
     is_system_mode: bool,
     base_path: Option<String>,
@@ -257,8 +254,8 @@ pub fn combined_json(
     let build = standard_output(
         input_files,
         libraries,
-        solc.to_owned(),
-        optimize,
+        solc,
+        optimizer_settings,
         force_evmla,
         is_system_mode,
         base_path,